fix: a bug fix
The user was able to open more than 2 cards at one time,solving this bug
by removing events handler from all elements,after clicking 2 cards,and
add them back when the 2 cards are closed.

Function are used to restart the game instead of reload the game,and
change the class of the button from restart to paly-again.